Move SceneModel to _model package and split it a bit
