+++ silx/gui/plot3d/_model/items.py
@@ -23,17 +23,16 @@
 #
 # ###########################################################################*/
 """
-This package implements the SceneWidget model
+This module provides base classes to implement models for 3D scene content
 """
 
 from __future__ import absolute_import, division
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "05/12/2017"
+__date__ = "11/01/2018"
 
 
-import collections
 import functools
 import weakref
 
@@ -41,292 +40,14 @@ import numpy
 
 from silx.third_party import six
 
-from ...utils.weakref import WeakMethodProxy
-from .._utils import convertArrayToQImage
-from ..plot.Colormap import preferredColormaps
-from .. import qt, icons
-from . import items
-from .items.volume import Isosurface
+from ..._utils import convertArrayToQImage
+from ...plot.Colormap import preferredColormaps
+from ... import qt, icons
+from .. import items
+from ..items.volume import Isosurface
 
 
-class BaseRow(qt.QObject):
-    """Base class for rows of the tree model.
-
-    The root node parent MUST be set to the QAbstractItemModel it belongs to.
-    By default item is enabled.
-
-    :param children: Iterable of BaseRow to start with (not signaled)
-    """
-
-    def __init__(self, children=()):
-        super(BaseRow, self).__init__()
-        self.__children = []
-        for row in children:
-            assert isinstance(row, BaseRow)
-            row.setParent(self)
-            self.__children.append(row)
-        self.__flags = collections.defaultdict(lambda: qt.Qt.ItemIsEnabled)
-
-    def model(self):
-        """Return the model this node belongs to or None if not in a model.
-
-        :rtype: Union[QAbstractItemModel, None]
-        """
-        parent = self.parent()
-        if isinstance(parent, BaseRow):
-            return self.parent().model()
-        elif parent is None:
-            return None
-        else:
-            assert isinstance(parent, qt.QAbstractItemModel)
-            return parent
-
-    def index(self, column=0):
-        """Return corresponding index in the model or None if not in a model.
-
-        :param int column: The column to make the index for
-        :rtype: Union[QModelIndex, None]
-        """
-        parent = self.parent()
-        model = self.model()
-
-        if model is None:  # Not in a model
-            return None
-        elif parent is model:  # Root node
-            return qt.QModelIndex()
-        else:
-            index = parent.index()
-            row = parent.children().index(self)
-            return model.index(row, column, index)
-
-    def columnCount(self):
-        """Returns number of columns (default: 2)
-
-        :rtype: int
-        """
-        return 2
-
-    def children(self):
-        """Returns the list of children nodes
-
-        :rtype: tuple of Node
-        """
-        return tuple(self.__children)
-
-    def rowCount(self):
-        """Returns number of rows
-
-        :rtype: int
-        """
-        return len(self.__children)
-
-    def addRow(self, row, index=None):
-        """Add a node to the children
-
-        :param BaseRow row: The node to add
-        :param int index: The index at which to insert it or
-                          None to append
-        """
-        if index is None:
-            index = self.rowCount()
-        assert index <= self.rowCount()
-
-        model = self.model()
-
-        if model is not None:
-            parent = self.index()
-            model.beginInsertRows(parent, index, index)
-
-        self.__children.insert(index, row)
-        row.setParent(self)
-
-        if model is not None:
-            model.endInsertRows()
-
-    def removeRow(self, row):
-        """Remove a row from the children list.
-
-        It removes either a node or a row index.
-
-        :param row: BaseRow object or index of row to remove
-        :type row: Union[BaseRow, int]
-        """
-        if isinstance(row, BaseRow):
-            row = self.__children.index(row)
-        else:
-            row = int(row)
-        assert row < self.rowCount()
-
-        model = self.model()
-
-        if model is not None:
-            index = self.index()
-            model.beginRemoveRows(index, row, row)
-
-        node = self.__children.pop(row)
-        node.setParent(None)
-
-        if model is not None:
-            model.endRemoveRows()
-
-    def data(self, column, role):
-        """Returns data for given column and role
-
-        :param int column: Column index for this row
-        :param int role: The role to get
-        :return: Corresponding data (Default: None)
-        """
-        return None
-
-    def setData(self, column, value, role):
-        """Set data for given column and role
-
-        :param int column: Column index for this row
-        :param value: The data to set
-        :param int role: The role to set
-        :return: True on success, False on failure
-        :rtype: bool
-        """
-        return False
-
-    def setFlags(self, flags, column=None):
-        """Set the static flags to return.
-
-        Default is ItemIsEnabled for all columns.
-
-        :param int column: The column for which to set the flags
-        :param flags: Item flags
-        """
-        if column is None:
-            self.__flags = collections.defaultdict(lambda: flags)
-        else:
-            self.__flags[column] = flags
-
-    def flags(self, column):
-        """Returns flags for given column
-
-        :rtype: int
-        """
-        return self.__flags[column]
-
-
-class StaticRow(BaseRow):
-    """Row with static data.
-
-    :param tuple display: List of data for DisplayRole for each column
-    :param dict roles: Optional mapping of roles to list of data.
-    :param children: Iterable of BaseRow to start with (not signaled)
-    """
-
-    def __init__(self, display=('', None), roles=None, children=()):
-        super(StaticRow, self).__init__(children)
-        self._dataByRoles = {} if roles is None else roles
-        self._dataByRoles[qt.Qt.DisplayRole] = display
-
-    def data(self, column, role):
-        if role in self._dataByRoles:
-            data = self._dataByRoles[role]
-            if column < len(data):
-                return data[column]
-        return super(StaticRow, self).data(column, role)
-
-    def columnCount(self):
-        return len(self._dataByRoles[qt.Qt.DisplayRole])
-
-
-class ProxyRow(BaseRow):
-    """Provides a node to proxy a data accessible through functions.
-
-    Warning: Only weak reference are kept on fget and fset.
-
-    :param str name: The name of this node
-    :param callable fget: A callable returning the data
-    :param callable fset:
-        An optional callable setting the data with data as a single argument.
-    :param notify:
-        An optional signal emitted when data has changed.
-    :param callable toModelData:
-        An optional callable to convert from fget
-        callable to data returned by the model.
-    :param callable fromModelData:
-        An optional callable converting data provided to the model to
-        data for fset.
-    :param editorHint: Data to provide as UserRole for editor selection/setup
-    """
-
-    def __init__(self,
-                 name='',
-                 fget=None,
-                 fset=None,
-                 notify=None,
-                 toModelData=None,
-                 fromModelData=None,
-                 editorHint=None):
-
-        super(ProxyRow, self).__init__()
-        self.__name = name
-        self.__editorHint = editorHint
-
-        assert fget is not None
-        self._fget = WeakMethodProxy(fget)
-        self._fset = WeakMethodProxy(fset) if fset is not None else None
-        if fset is not None:
-            self.setFlags(self.flags(1) | qt.Qt.ItemIsEditable, 1)
-        self._toModelData = toModelData
-        self._fromModelData = fromModelData
-
-        if notify is not None:
-            notify.connect(self._notified)  # TODO support sigItemChanged flags
-
-    def _notified(self, *args, **kwargs):
-        """Send update to the model upon signal notifications"""
-        index = self.index(column=1)
-        model = self.model()
-        if model is not None:
-            model.dataChanged.emit(index, index)
-
-    def data(self, column, role):
-        if column == 0:
-            if role == qt.Qt.DisplayRole:
-                return self.__name
-
-        elif column == 1:
-            if role == qt.Qt.UserRole:  # EditorHint
-                return self.__editorHint
-            elif role == qt.Qt.DisplayRole or (role == qt.Qt.EditRole and
-                                               self._fset is not None):
-                data = self._fget()
-                if self._toModelData is not None:
-                    data = self._toModelData(data)
-                return data
-
-        return super(ProxyRow, self).data(column, role)
-
-    def setData(self, column, value, role):
-        if role == qt.Qt.EditRole and self._fset is not None:
-            if self._fromModelData is not None:
-                value = self._fromModelData(value)
-            self._fset(value)
-            return True
-
-        return super(ProxyRow, self).setData(column, value, role)
-
-
-class ColorProxyRow(ProxyRow):
-    """Provides a proxy to a QColor property.
-
-    The color is returned through the decorative role.
-
-    See :class:`ProxyRow`
-    """
-
-    def data(self, column, role):
-        if column == 1:  # Show color as decoration, not text
-            if role == qt.Qt.DisplayRole:
-                return None
-            if role == qt.Qt.DecorationRole:
-                role = qt.Qt.DisplayRole
-        return super(ColorProxyRow, self).data(column, role)
+from .core import AngleDegreeRow, BaseRow, ColorProxyRow, ProxyRow, StaticRow
 
 
 class _DirectionalLightProxy(qt.QObject):
@@ -524,22 +245,6 @@ class Item3DRow(StaticRow):
             else:
                 return False
         return super(Item3DRow, self).setData(column, value, role)
-
-
-class AngleDegreeRow(ProxyRow):
-    """ProxyRow patching display of column 1 to add degree symbol
-
-    See :class:`ProxyRow`
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(AngleDegreeRow, self).__init__(*args, **kwargs)
-
-    def data(self, column, role):
-        if column == 1 and role == qt.Qt.DisplayRole:
-            return u'%g°' % super(AngleDegreeRow, self).data(column, role)
-        else:
-            return super(AngleDegreeRow, self).data(column, role)
 
 
 class DataItem3DTransformRow(StaticRow):
@@ -1452,119 +1157,3 @@ def nodeFromItem(item, name=None):
                 node.addRow(PlaneRow(item))
 
     return node
-
-
-class Root(BaseRow):
-    """Root node of :class:`SceneWidget` parameters.
-
-    It has two children:
-    - Settings
-    - Scene group
-    """
-
-    def __init__(self, model, sceneWidget):
-        super(Root, self).__init__()
-        self._sceneWidget = weakref.ref(sceneWidget)
-        self.setParent(model)  # Needed for Root
-
-    def children(self):
-        sceneWidget = self._sceneWidget()
-        if sceneWidget is None:
-            return ()
-        else:
-            return super(Root, self).children()
-
-
-class SceneModel(qt.QAbstractItemModel):
-    """Model of a :class:`SceneWidget`.
-
-    :param SceneWidget parent: The SceneWidget this model represents.
-    """
-
-    def __init__(self, parent):
-        self._sceneWidget = weakref.ref(parent)
-
-        super(SceneModel, self).__init__(parent)
-        self._root = Root(self, parent)
-        self._root.addRow(Settings(parent))
-        self._root.addRow(nodeFromItem(parent.getSceneGroup(), name='Data'))
-
-    def sceneWidget(self):
-        """Returns the :class:`SceneWidget` this model represents.
-
-        In case the widget has already been deleted, it returns None
-
-        :rtype: SceneWidget
-        """
-        return self._sceneWidget()
-
-    def _itemFromIndex(self, index):
-        """Returns the corresponding :class:`Node` or :class:`Item3D`.
-
-        :param QModelIndex index:
-        :rtype: Node or Item3D
-        """
-        return index.internalPointer() if index.isValid() else self._root
-
-    def index(self, row, column, parent=qt.QModelIndex()):
-        """See :meth:`QAbstractItemModel.index`"""
-        if column >= self.columnCount(parent) or row >= self.rowCount(parent):
-            return qt.QModelIndex()
-
-        item = self._itemFromIndex(parent)
-        return self.createIndex(row, column, item.children()[row])
-
-    def parent(self, index):
-        """See :meth:`QAbstractItemModel.parent`"""
-        if not index.isValid():
-            return qt.QModelIndex()
-
-        item = self._itemFromIndex(index)
-        parent = item.parent()
-
-        ancestor = parent.parent()
-
-        if ancestor is not self:  # root node
-            children = ancestor.children()
-            row = children.index(parent)
-            return self.createIndex(row, 0, parent)
-
-        return qt.QModelIndex()
-
-    def rowCount(self, parent=qt.QModelIndex()):
-        """See :meth:`QAbstractItemModel.rowCount`"""
-        item = self._itemFromIndex(parent)
-        return item.rowCount()
-
-    def columnCount(self, parent=qt.QModelIndex()):
-        """See :meth:`QAbstractItemModel.columnCount`"""
-        item = self._itemFromIndex(parent)
-        return item.columnCount()
-
-    def data(self, index, role=qt.Qt.DisplayRole):
-        """See :meth:`QAbstractItemModel.data`"""
-        item = self._itemFromIndex(index)
-        column = index.column()
-        return item.data(column, role)
-
-    def setData(self, index, value, role=qt.Qt.EditRole):
-        """See :meth:`QAbstractItemModel.setData`"""
-        item = self._itemFromIndex(index)
-        column = index.column()
-        if item.setData(column, value, role):
-            self.dataChanged.emit(index, index)
-            return True
-        return False
-
-    def flags(self, index):
-        """See :meth:`QAbstractItemModel.flags`"""
-        item = self._itemFromIndex(index)
-        column = index.column()
-        return item.flags(column)
-
-    def headerData(self, section, orientation, role=qt.Qt.DisplayRole):
-        """See :meth:`QAbstractItemModel.headerData`"""
-        if orientation == qt.Qt.Horizontal and role == qt.Qt.DisplayRole:
-            return 'Item' if section == 0 else 'Value'
-        else:
-            return None
